# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Wisuda Headless API Engine",
    "description": "API Engine untuk Platform Wisuda Photography Agency. Mendukung otentikasi JWT, Session Cookie, & API Key untuk integrasi multi-client.",
    "version": "1.3.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Development Server"
    },
    {
      "url": "https://api.wisuda.id",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root API Info",
        "responses": {
          "200": {
            "description": "Informasi status & endpoint backend API"
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "summary": "Health Check Status",
        "responses": {
          "200": {
            "description": "Database & server status OK"
          }
        }
      }
    },
    "/api/admin/login": {
      "post": {
        "summary": "Admin / User Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil, mengembalikan JWT Token & User Profile"
          },
          "401": {
            "description": "Kredensial tidak valid"
          }
        }
      }
    },
    "/api/admin/api-keys": {
      "get": {
        "summary": "List API Keys",
        "security": [{ "BearerAuth": [] }, { "ApiKeyAuth": [] }],
        "responses": {
          "200": {
            "description": "Daftar API Key aktif"
          }
        }
      },
      "post": {
        "summary": "Buat API Key Baru",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "scopes": { "type": "string", "example": "read,write" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API Key berhasil dibuat"
          }
        }
      }
    }
  }
}